Remove file with rrd data when a plugin is closed.
Removing the file is a way to signal the RRD daemon that the plugin is
closed. Otherwise some other mechanism would be needed. This requires
storing the path in the internal data structure.

The test cases have been adapted. Previously tests relied on the fact
that files written by a plugin would exist beyond the lifetime of the
plugin. They now test that these file are removed.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>